requirements.txt
fastapi
uvicorn>=0.28.0
pydantic>=2.10.0
pyyaml>=6.0.1
cel-python